requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.5
djangorestframework==3.10.3
idna==2.8
importlib-metadata==0.19
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
more-itertools==7.2.0
packaging==19.1
pika==1.1.0
pluggy==0.12.0
py==1.8.0
PyJWT==1.7.1
pyparsing==2.4.2
pytest==5.1.2
python-consul==1.1.0
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
wcwidth==0.1.7
zipp==0.6.0
pytest-cov==2.7.1
openapi_codec==1.3.2
dulwich==0.19.13
cryptography==2.8
